% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrostab.R
\docType{package}
\name{agrostab}
\alias{agrostab}
\alias{agrostab-package}
\title{Stability Analysis for Agricultural Research}
\description{
The \code{agrostab} package provides functionalities to perform stability analysis in plant breeding.
The package includes statistical procedures to identify stable genotypes under diverse environments.
}
\author{
{
Anna Cheshkova <cheshanf@gmail.com>
}
}
