% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_data.R
\docType{data}
\name{exp_data}
\alias{exp_data}
\title{Experimental data for stability analysis}
\format{A \code{data.frame} 126 obs.  of 4 variables.}
\usage{
data(exp_data)
}
\description{
Data obtained from  the agrotechnical experiments carried out in 2009-2011 to evaluate grain yield of seven Siberian common winter wheat cultivars.
}
\details{
\itemize{
       \item{env} Environment
       \item{gen} Genotype
       \item{rep} Replicate
       \item{yield} Yield Response
       }
}
\examples{
data(exp_data)

}
\references{
Siberian Research Institute of Plant Growing and Breeding - Branch of the Institute of Cytology and Genetics,
 Krasnoobsk, Novosibirsk region, Russia
}
\keyword{datasets}
